+++ antflow-engine/src/main/java/org/openoa/engine/bpmnconf/confentity/BpmnConf.java
@@ -88,7 +88,7 @@ public class BpmnConf {
      * business party mark
      */
     @TableField("business_party_id")
-    private Integer businessPartyId;
+    private Long businessPartyId;
     @TableField("extra_flags")
     private Integer extraFlags;
     /**
