+++ antflow-engine/src/main/java/org/openoa/engine/bpmnconf/service/impl/OutSideBpmBusinessPartyServiceImpl.java
@@ -224,7 +224,7 @@ public class OutSideBpmBusinessPartyServiceImpl extends ServiceImpl<OutSideBpmBu
      * @param id
      * @return
      */
-    public String getBusinessPartyMarkById(Integer id) {
+    public String getBusinessPartyMarkById(Long id) {
         return outSideBpmBusinessPartyMapper.getBusinessPartyMarkById(id);
     }
